Transpose in TBN
Previously, transpose was removed in TBN for fragment shader. Before releasing assignment, it was thought that indecision was present about use of TBN in vert or frag shader -- the former requiring the inverse TBN (i.e., transpose given orthonormal basis) to convert the light direction to the local tangent space of a model's triangles. However, the original transpose was in the frag shader, not because of uncertainty in using the inverse TBN in the vert shader (lower performance cost), but instead because the constructor args of a 3x3 matrix put the Tangent, Bitangent, and Normal vectors in the wrong col/row order. Therefore, the original transpose was not for the pseudoinverse, but instead to correct the col/row ordering.